move adapters to their own module
