+++ http_check/datadog_checks/http_check/http_check.py
@@ -1,10 +1,8 @@
 # (C) Datadog, Inc. 2018
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
-
 from __future__ import unicode_literals
 
-# stdlib
 from datetime import datetime
 import _strptime # noqa
 import os.path
@@ -15,19 +13,16 @@ import time
 import warnings
 from urlparse import urlparse
 
-# 3rd party
 import requests
-from requests.adapters import HTTPAdapter
-from requests.packages import urllib3
-from requests.packages.urllib3.util import ssl_
-from requests.packages.urllib3.exceptions import InsecureRequestWarning, SecurityWarning
-from requests.packages.urllib3.packages.ssl_match_hostname import match_hostname
+from requests.packages.urllib3.exceptions import InsecureRequestWarning
 from requests_ntlm import HttpNtlmAuth
 
-# project
 from datadog_checks.checks import NetworkCheck, Status
 from datadog_checks.config import _is_affirmative
 from datadog_checks.utils.headers import headers as agent_headers
+
+from .adapters import WeakCiphersAdapter, WeakCiphersHTTPSConnection
+
 
 DEFAULT_EXPECTED_CODE = "(1|2|3)\d\d"
 DEFAULT_EXPIRE_DAYS_WARNING = 14
@@ -37,97 +32,6 @@ DEFAULT_EXPIRE_CRITICAL = DEFAULT_EXPIRE_DAYS_CRITICAL * 24 * 3600
 CONTENT_LENGTH = 200
 
 DATA_METHODS = ['POST', 'PUT', 'DELETE', 'PATCH']
-
-
-class WeakCiphersHTTPSConnection(urllib3.connection.VerifiedHTTPSConnection):
-
-    SUPPORTED_CIPHERS = (
-        'ECDH+AESGCM:DH+AESGCM:ECDH+AES256:DH+AES256:ECDH+AES128:DH+AES:'
-        'ECDH+HIGH:DH+HIGH:ECDH+3DES:DH+3DES:RSA+AESGCM:RSA+AES:RSA+HIGH:'
-        'RSA+3DES:ECDH+RC4:DH+RC4:RSA+RC4:!aNULL:!eNULL:!EXP:-MD5:RSA+RC4+MD5'
-    )
-
-    def __init__(self, host, port, ciphers=None, **kwargs):
-        self.ciphers = ciphers if ciphers is not None else self.SUPPORTED_CIPHERS
-        super(WeakCiphersHTTPSConnection, self).__init__(host, port, **kwargs)
-
-    def connect(self):
-        # Add certificate verification
-        conn = self._new_conn()
-
-        resolved_cert_reqs = ssl_.resolve_cert_reqs(self.cert_reqs)
-        resolved_ssl_version = ssl_.resolve_ssl_version(self.ssl_version)
-
-        hostname = self.host
-        if getattr(self, '_tunnel_host', None):
-            # _tunnel_host was added in Python 2.6.3
-            # (See:
-            # http://hg.python.org/cpython/rev/0f57b30a152f)
-            #
-            # However this check is still necessary in 2.7.x
-
-            self.sock = conn
-            # Calls self._set_hostport(), so self.host is
-            # self._tunnel_host below.
-            self._tunnel()
-            # Mark this connection as not reusable
-            self.auto_open = 0
-
-            # Override the host with the one we're requesting data from.
-            hostname = self._tunnel_host
-
-        # Wrap socket using verification with the root certs in trusted_root_certs
-        self.sock = ssl_.ssl_wrap_socket(conn, self.key_file, self.cert_file,
-                                         cert_reqs=resolved_cert_reqs,
-                                         ca_certs=self.ca_certs,
-                                         server_hostname=hostname,
-                                         ssl_version=resolved_ssl_version,
-                                         ciphers=self.ciphers)
-
-        if self.assert_fingerprint:
-            ssl_.assert_fingerprint(self.sock.getpeercert(binary_form=True), self.assert_fingerprint)
-        elif resolved_cert_reqs != ssl.CERT_NONE \
-                and self.assert_hostname is not False:
-            cert = self.sock.getpeercert()
-            if not cert.get('subjectAltName', ()):
-                warnings.warn((
-                    'Certificate has no `subjectAltName`, falling back to check for a `commonName` for now. '
-                    'This feature is being removed by major browsers and deprecated by RFC 2818. '
-                    '(See https://github.com/shazow/urllib3/issues/497 for details.)'),
-                    SecurityWarning
-                )
-            match_hostname(cert, self.assert_hostname or hostname)
-
-        self.is_verified = (resolved_cert_reqs == ssl.CERT_REQUIRED or self.assert_fingerprint is not None)
-
-
-class WeakCiphersHTTPSConnectionPool(urllib3.connectionpool.HTTPSConnectionPool):
-
-    ConnectionCls = WeakCiphersHTTPSConnection
-
-
-class WeakCiphersPoolManager(urllib3.poolmanager.PoolManager):
-
-    def _new_pool(self, scheme, host, port):
-        if scheme == 'https':
-            return WeakCiphersHTTPSConnectionPool(host, port, **(self.connection_pool_kw))
-        return super(WeakCiphersPoolManager, self)._new_pool(scheme, host, port)
-
-
-class WeakCiphersAdapter(HTTPAdapter):
-    """"Transport adapter" that allows us to use TLS_RSA_WITH_RC4_128_MD5."""
-
-    def init_poolmanager(self, connections, maxsize, block=False, **pool_kwargs):
-        # Rewrite of the
-        # requests.adapters.HTTPAdapter.init_poolmanager method
-        # to use WeakCiphersPoolManager instead of
-        # urllib3's PoolManager
-        self._pool_connections = connections
-        self._pool_maxsize = maxsize
-        self._pool_block = block
-
-        self.poolmanager = WeakCiphersPoolManager(num_pools=connections,
-                                                  maxsize=maxsize, block=block, strict=True, **pool_kwargs)
 
 
 def get_ca_certs_path():
